Check file fully working
Could still do with a bit of tidying - better to be consistent and convert all H5 refs to paths in code. Only bit to keep as refs is bit to find unique ones as this should be possible without dereferencing, although at present that is exactly how it is done, followed by identifying unique paths and converting back to refs!